From wbrouw@atnf.csiro.au Thu May 22 17:29:36 1997 To: Neil Killeen <nkilleen@atnf.csiro.au> cc: wbrouw@atnf.csiro.au Subject: Re: pureatria Date: Thu, 22 May 1997 17:29:34 +1000 From: "Wim Brouw" <wbrouw@atnf.csiro.au> Content-Length: 568
 >
 > gee wim. i even i know the cost codes
 > that i manage.,
 >
That's why you are maging..
 > i am having fun with virtual functions and
 > abstarct base classes today.  i am
 > tryihng to add one default vaklue to one
 > function in one class
I gave up on default values if they are not a standard type. Especially SUN
ntv
does not like it; accepts only constants; but then the order of statics kills
you..
 > !! wheeeeeeeeeewhere are you when i need you !
Where it is wet and cool of course; discussing nice things like Tables, Mns,
and even some fun.

Wim